refactor: reuse MCP client for tools and prompts
Refactors the MCP client handling to reuse the same client for fetching both tools and prompts from each server. This improves efficiency by:
- Creating a single client per server
- Fetching both tools and prompts in one pass
- Using a shared initialization process
- Properly managing client lifecycle
